% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{add.geno}
\alias{add.geno}
\title{Sum haplotypes to get genotypes}
\usage{
add.geno(a, b)
}
\arguments{
\item{a}{vector with haplotypes 1}

\item{b}{vector with haplotypes}
}
\value{
vector  with the sum of the haplotype pair (genotype)
add.geno()
}
\description{
This function allows you to sum a pair of haplotypes (strings) and return a string with the sum
}
\keyword{haplotype}
\keyword{pairs}
\keyword{sum}
