% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{add.prior}
\alias{add.prior}
\title{Add any mixture of Gaussians to stan inputs}
\usage{
add.prior(prior, l)
}
\arguments{
\item{prior}{named list: mean= vector with the mean of Gaussians, sd= vector with Gaussians sd for eQTL effect prior, mix=vector with mixing proportions}

\item{l}{list with stan input already made to append prior info}
}
\value{
list with addedd prior info
add.prior()
}
\description{
This functions allows adding any Gaussian (or mixture) to be used as prior for eQTL effect when runing stan model
}
