% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_noGT_2T.R
\name{aux.in}
\alias{aux.in}
\title{Third aux function to prepare inputs to run baseqtl without rsnp GT: prepare stan inputs wrap}
\usage{
aux.in(
  gene,
  ai = NULL,
  case,
  rp.f,
  rp.r,
  f.ase,
  counts.g,
  covariates,
  min.ase = 5,
  min.ase.n = 5,
  info = 0.3,
  snps.ex,
  prefix = NULL,
  out = ".",
  prior = NULL,
  mc.cores = getOption("mc.cores", 1)
)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{ai}{list with estimates for allelic imbalance per skin, defaults to NULL}

\item{case, }{list with data table with ASE counts per skin}

\item{rp.f}{list with reference panel haps for fSNPs per skin}

\item{rp.r}{ref panel haps for rSNPs}

\item{f.ase}{list with data table with genotypes for fSNPs per skin}

\item{counts.g}{list with data table with total counts for gene per skin}

\item{covariates}{per skin}

\item{min.ase}{minimun number of ASE counts for an individual in order to be included, defaults to 5}

\item{min.ase.n}{minimun number individuals with the minimun of ASE counts, defaults to 5.}

\item{info}{numeric cut-off for var(E(G))/var(G), var(E(G)) is the expected variance for input and var(G) for reference panel, similar to info score in impute2, defaults to 0.3. rsnps with lower info wont be run by stan.}

\item{snps.ex}{data table with rsnps to exclude from running}

\item{prefix}{character with prefix to add for saving files, defaults to NULL}

\item{out}{path to save outputs, default to current directory}

\item{prior}{named list: mean= vector with the mean of Gaussians, sd= vector with Gaussians sd for eQTL effect prior, mix=vector with mixing proportions. Defaults to NULL, mixture of 3 components with mean (0,0,0) ;sd  c( 0.0436992,  0.3492696, 0.4920049); and mixing proportions  c(0.955, 2*0.015, 0.015).}

\item{mc.cores}{The number of parallel cores to use.}
}
\value{
list with stan input
}
\description{
This function allows you to prepare stan inputs final stages
}
