% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_noGT.R
\name{aux.in1}
\alias{aux.in1}
\title{First aux function to prepare inputs to run baseqtl with or without known rsnp GT: check inputs}
\usage{
aux.in1(
  gene,
  chr,
  snps = 5 * 10^5,
  counts.f,
  covariates = 1,
  additional_cov = NULL,
  e.snps,
  u.esnps = NULL,
  gene.coord,
  vcf,
  sample.file = NULL,
  le.file,
  h.file,
  population = c("EUR", "AFR", "AMR", "EAS", "SAS", "ALL"),
  maf = NULL,
  nhets = NULL,
  min.ase = 5,
  min.ase.het = NULL,
  min.ase.snp = NULL,
  min.ase.n = NULL,
  tag.threshold = 0.9,
  info = NULL,
  out = ".",
  model = NULL,
  prob = NULL,
  AI_estimate = NULL,
  pretotalReads = 100
)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{chr}{chromosome where the gene is, example chr=22}

\item{snps}{either cis-window or character vector with pos:ref:alt allele for each snp, defaults to cis-window}

\item{counts.f}{path to file with filtered counts: rows genes, first col gene_id followed by samples, prepared in inputs.R}

\item{covariates}{path to matrix of covariates prepared in inputs.R, if no covariates, covariates =1, default}

\item{additional_cov}{full name to file with first column sample names and additional columns gene independent covariates, defaults to NULL}

\item{e.snps}{path to file listing exonic snps for the chromosome where the gene is, prepared in input.R}

\item{u.esnps}{whether to use unique exonic snps per gene, defaults to NULL when it is not necessary if strand info is known}

\item{gene.coord}{path to file listing gene coordinates and exons, prepared in input.R}

\item{vcf}{path to vcf file with ASE and GT for the chromosome where the gene is}

\item{sample.file}{sample file for the chromosome of interest for the reference panel (sample description), to be used if ex.fsnp test is required and populatios is not the whole reference panel}

\item{le.file}{path to gz legend file (legend.gz) for the chromosome of interest for the reference panel (snp description)}

\item{h.file}{path to gz haplotpe file for the chromosome of interest for the reference panel (haplotypes for all samples in reference panel)}

\item{population}{ethnicity to get EAF for rsnp: AFR AMR EAS EUR SAS ALL, defaults to EUR. It is also used to exclude fSNPs if frequency is different from reference panel when using no GT and selecting ex.fsnps argument, except if population=ALL}

\item{maf}{cut-off for maf, defaults to NULL as only used with noGT}

\item{nhets}{minimun number of het individuals in order to run the minumn model (NB only), defaults to 5}

\item{min.ase}{minimun number of ASE counts for an individual in order to be included, defaults to 5}

\item{min.ase.het}{minimun number of het individuals with the minimun of ASE counts in order to run the ASE side of the model, defaults to 5.}

\item{tag.threshold}{numeric with r2 threshold (0-1) for grouping snps to reduce the number of running tests, to disable use "no"}

\item{info}{numeric cut-off for var(E(G))/var(G), var(E(G)) is the expected variance for input and var(G) for reference panel, similar to info score in impute2, defaults to 0.3. rsnps with lower info wont be run by stan.}

\item{out}{path to save outputs, default to current directory}

\item{model}{GT: whether to run NB-ASE (full model negative binomial and allele specific counts),NB (negative binomial only) or both (NB-ASE and NB for those associations with no ASE information), noGT NULL, defaults to NULL}

\item{prob}{number p∈(0,1) indicating the desired probability mass to include in the intervals, defaults to 0.99 and 0.95 quantiles}

\item{AI_estimate}{full name to data table with AI estimates for reference panel bias for fSNPs, defaults to NULL}

\item{pretotalReads}{numeric indicating a cut-off for total initial reads to consider AI estimates, defaults to 100}

\item{prefix}{optional prefix for saving tables, if NULL gene_id.eqtl will be used}
}
\value{
list with counts, covariates and probs argument to prepare inputs
aux.in1()
}
\description{
This function allows you to check arguments with and without GT
}
