% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_noGT.R
\name{aux.in2}
\alias{aux.in2}
\title{Second aux function to prepare inputs to run baseqtl with or without known rsnp GT: ASE counts mapping unique fSNPs}
\usage{
aux.in2(gene, u.esnps, case, ex.fsnp = NULL, ai = NULL)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{u.esnps}{whether to use unique exonic snps per gene, defaults to NULL when it is not necessary if strand info is known}

\item{case, }{data table with ASE counts}

\item{ex.fsnp}{character vector with id of fsnps to exclude}

\item{ai}{data table with allelic imbalance estimates}
}
\value{
data table with filtered ASE counts
aux.in2()
}
\description{
This function allows you to check arguments with and without GT
}
