% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_noGT.R
\name{aux.in3}
\alias{aux.in3}
\title{Third aux function to prepare inputs to run baseqtl without known rsnp GT: prepare stan inputs wrap}
\usage{
aux.in3(
  gene,
  ai = NULL,
  case,
  rp.f,
  rp.r,
  f.ase,
  counts.g,
  covariates,
  min.ase = 5,
  min.ase.n = 5,
  info = 0.3,
  snps.ex,
  prefix = NULL,
  out = "."
)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{ai}{estimates for allelic imbalance, defaults to NULL}

\item{case, }{data table with ASE counts}

\item{rp.f}{reference panel haps for fSNPs}

\item{rp.r}{ref panel haps for rSNPs}

\item{f.ase}{data table with genotypes for fSNPs}

\item{counts.g}{data table with total counts for gene}

\item{covariates}{path to matrix of covariates prepared in inputs.R. If using gc correction (each gene diffrent value), the matrix has rownames= genes and cols=samples plus extra columns if other covariates are added. If only using lib size or gene independent covariates, rows are samples and columns are covariates. If no covariates, covariates =1, default}

\item{min.ase}{minimun number of ASE counts for an individual in order to be included, defaults to 5}

\item{min.ase.n}{minimun number individuals with the minimun of ASE counts, defaults to 5.}

\item{info}{numeric cut-off for var(E(G))/var(G), var(E(G)) is the expected variance for input and var(G) for reference panel, similar to info score in impute2, defaults to 0.3. rsnps with lower info wont be run by stan.}

\item{snps.ex}{data table with rsnps to exclude from running}

\item{prefix}{character with prefix to add for saving files, defaults to NULL}

\item{out}{path to save outputs, default to current directory}
}
\value{
list with stan input
aux.in3()
}
\description{
This function allows you to prepare stan inputs final stages
}
