% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{aux.sort}
\alias{aux.sort}
\title{Aux function to sort ase according by ascending genotype order (based on NB genotype)}
\usage{
aux.sort(a, b)
}
\arguments{
\item{a}{list to sort}

\item{ind}{list with indexes}
}
\value{
vector to input for stan

aux.sort()
}
\description{
Aux function to sort ase according by ascending genotype order (based on NB genotype)
}
\keyword{Gi=g}
\keyword{ase}
\keyword{by}
\keyword{sort}
