% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{aux.stan.ase}
\alias{aux.stan.ase}
\title{aux function to prepare inputs for allelic imbalance estimates, wrapper to avoid code repetition}
\usage{
aux.stan.ase(mat, g.rsnp, n.mat)
}
\arguments{
\item{mat}{matrix with allelic imbalance or variance allelic imbalance estimates}

\item{g.rsnp}{object created in stan.ase containing genotypes compatible with reference panel}

\item{n.mat}{matrix with hap for fsnps and ase counts, output from fsnp.prep or stan.fsnp.noGT}
}
\value{
list of stan input: list with AI reference bias estimates or variance estimates per individual

aux.stan.ase()
}
\description{
aux function to prepare inputs for allelic imbalance estimates, wrapper to avoid code repetition
}
