% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseqtl-package.R
\docType{package}
\name{baseqtl-package}
\alias{baseqtl-package}
\alias{baseqtl}
\title{The 'baseqtl' package.}
\description{
Bayesian allele specific expression for eqtl analysis
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
