% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_noGT.R
\name{baseqtl.nogt.in}
\alias{baseqtl.nogt.in}
\title{Prepare inputs for baseqtl with unknown rsnp GT and missing values for GT fsnps with reference panel bias correction}
\usage{
baseqtl.nogt.in(
  gene,
  chr,
  snps = 5 * 10^5,
  counts.f,
  covariates = 1,
  e.snps,
  u.esnps = NULL,
  gene.coord,
  vcf,
  sample.file = NULL,
  le.file,
  h.file,
  population = c("EUR", "AFR", "AMR", "EAS", "SAS", "ALL"),
  maf = 0.05,
  min.ase = 5,
  min.ase.snp = 5,
  min.ase.n = 5,
  tag.threshold = 0.9,
  info = 0.3,
  out = ".",
  model = NULL,
  prefix = NULL,
  ex.fsnp = NULL,
  prob = NULL,
  AI_estimate = NULL,
  pretotalReads = 100
)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{chr}{chromosome where the gene is, example chr=22}

\item{snps}{either cis-window or character vector with pos:ref:alt allele for each snp, defaults to cis-window}

\item{counts.f}{path to file with filtered counts: rows genes, first col gene_id followed by samples, prepared in inputs.R}

\item{covariates}{path to matrix of covariates prepared in inputs.R. If using gc correction (each gene diffrent value), the matrix has rownames= genes and cols=samples plus extra columns if other covariates are added. If only using lib size or gene independent covariates, rows are samples and columns are covariates. If no covariates, covariates =1, default}

\item{e.snps}{path to file listing exonic snps for the chromosome where the gene is, prepared in input.R}

\item{u.esnps}{whether to use unique exonic snps per gene, defaults to NULL when it is not necessary if strand info is known}

\item{gene.coord}{path to file listing gene coordinates and exons, prepared in input.R}

\item{vcf}{path to vcf file with ASE and GT for the exonic snps for the chromosome where the gene is}

\item{sample.file}{sample file for the reference panel (sample description), to be used if ex.fsnp test is required and populatios is not the whole reference panel}

\item{le.file}{path to gz legend file (legend.gz) for the chromosome of interest for the reference panel (snp description)}

\item{h.file}{path to gz haplotpe file for the chromosome of interest for the reference panel (haplotypes for all samples in reference panel)}

\item{population}{ethnicity to set a cut-off for maf: AFR AMR EAS EUR SAS ALL, defaults to EUR}

\item{maf}{cut-off for maf, defaults to 0.05}

\item{min.ase}{minimun number of ASE counts for an individual in order to be included, defaults to 5}

\item{min.ase.snp}{minum number of ASE counts for a single snp to be considered, for a particular individual}

\item{min.ase.n}{minimun number individuals with the minimun of ASE counts, defaults to 5.}

\item{tag.threshold}{numeric with r2 threshold (0-1) for grouping snps to reduce the number of running tests, to disable use "no"}

\item{info}{numeric cut-off for var(E(G))/var(G), var(E(G)) is the expected variance for input and var(G) for reference panel, similar to info score in impute2, defaults to 0.3. rsnps with lower info wont be run by stan.}

\item{out}{path to save outputs, default to current directory}

\item{model}{GT: whether to run NB-ASE (full model negative binomial and allele specific counts),NB (negative binomial only) or both (NB-ASE and NB for those associations with no ASE information), noGT NULL, defaults to NULL}

\item{ex.fsnp, }{if character: vector with pos:ref:alt for fsnps to exclude, if numeric  p-value cut-off for fSNPs to exclude,  defaults to NULL}

\item{prob}{number p∈(0,1) indicating the desired probability mass to include in the intervals, defaults to 0.99 and 0.95 quantiles}

\item{AI_estimate}{full name to data table with AI estimates for reference panel bias for fSNPs, defaults to NULL}

\item{pretotalReads}{numeric indicating a cut-off for total initial reads to consider AI estimates, defaults to 100}
}
\value{
list with 1)c.ase and 2)stan.noGT object
baseqtl.nogt.in()
}
\description{
This function allows you to get inputs for baseqtl for one gene and multiple pre-selected snps.
}
