% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseqtl_noGT_2T.R
\name{baseqtl2T.nogt}
\alias{baseqtl2T.nogt}
\title{Run baseqtl for 2 treatments  in one model from different smples
This function allows you to run baseqtl for one gene and multiple pre-selected snps.}
\usage{
baseqtl2T.nogt(
  gene,
  chr,
  snps = 5 * 10^5,
  counts.f,
  covariates = 1,
  e.snps,
  u.esnps = NULL,
  gene.coord,
  vcf,
  sample.file = NULL,
  le.file,
  h.file,
  population = c("EUR", "AFR", "AMR", "EAS", "SAS", "ALL"),
  maf = 0.05,
  min.ase = 5,
  min.ase.snp = 5,
  min.ase.n = 5,
  tag.threshold = 0.9,
  info = 0.3,
  out = ".",
  prefix = NULL,
  model = NULL,
  prob = NULL,
  prior = NULL,
  ex.fsnp = 0.01,
  AI_estimate = NULL,
  pretotalReads = 100,
  treatments,
  fishjoin = "joint"
)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{chr}{chromosome where the gene is, example chr=22}

\item{snps}{either cis-window or character vector with pos:ref:alt allele for each snp, defaults to cis-window}

\item{counts.f}{path to files with filtered counts: rows genes, first col gene_id followed by samples, prepared in inputs.R}

\item{covariates}{path to matrix of covariates prepared in inputs.R, if no covariates, covariates =1, default. Same  order of treatments as counts.f}

\item{e.snps}{path to file listing exonic snps for the chromosome where the gene is, prepared in input.R}

\item{u.esnps}{whether to use unique exonic snps per gene, defaults to NULL when it is not necessary if strand info is known}

\item{gene.coord}{path to file listing gene coordinates and exons, prepared in input.R}

\item{vcf}{path to vcf files with ASE and GT for the exonic snps for the chromosome where the gene is, same order of treatment as counts.f and covariates
@param sample.file sample file for the reference panel (sample description), to be used if ex.fsnp test is required and population is not the whole reference panel}

\item{le.file}{path to gz legend file (legend.gz) for the chromosome of interest for the reference panel (snp description)}

\item{h.file}{path to gz haplotpe file for the chromosome of interest for the reference panel (haplotypes for all samples in reference panel)}

\item{population}{ethnicity to set a cut-off for maf: AFR AMR EAS EUR SAS ALL, defaults to EUR}

\item{maf}{cut-off for maf, defaults to 0.05}

\item{min.ase}{minimun number of ASE counts for an individual in order to be included, defaults to 5}

\item{min.ase.snp}{minum number of ASE counts for a single snp to be considered, for a particular individual}

\item{min.ase.n}{minimun number individuals with the minimun of ASE counts, defaults to 5.}

\item{tag.threshold}{numeric with r2 threshold (0-1) for grouping snps to reduce the number of running tests, to disable use "no"}

\item{info}{numeric cut-off for var(E(G))/var(G), var(E(G)) is the expected variance for input and var(G) for reference panel, similar to info score in impute2, defaults to 0.3. rsnps with lower info wont be run by stan.}

\item{out}{path to save outputs, default to current directory}

\item{prefix}{optional prefix for saving files, if NULL gene_id will be used}

\item{model}{compiled stanmodel object with stan model, defaults NULL which uses  built-in NB-ASE model. When AI_estimate is provided the model corrects for reference panel bias, otherwise id doesn't.}

\item{prob}{number p∈(0,1) indicating the desired probability mass to include in the intervals, defaults to 0.95 and 0.99 quantiles}

\item{prior}{named list: mean= vector with the mean of Gaussians, sd= vector with Gaussians sd for eQTL effect prior, mix=vector with mixing proportions. Defaults to NULL, mixture of 3 components with mean (0,0,0) ;sd  c( 0.0436992,  0.3492696, 0.4920049); and mixing proportions  c(0.9478807,0.05142192, 0.0006974015).}

\item{ex.fsnp}{if character: vector with pos:ref:alt for fsnps to exclude, if numeric  p-value cut-off for fSNPs to exclude based on fisher test for suspected genotype error based on comparing the proportion of hets in the sample and reference panel,  defaults to 0.01. If NULL no fSNP will be excluded.}

\item{AI_estimate}{full name to data table with AI estimates for reference panel bias for fSNPs, defaults to NULL}

\item{pretotalReads}{numeric indicating a cut-off for total initial reads to consider AI estimates, defaults to 100}

\item{treatments}{character vector with the 2 treatments (tissues, diseases) to study, sam order as counts.f, covariates and vcf}

\item{fishjoin}{whether to run Fisher test jointly in all samples or by treatment, defaults to jointly but  for QC purposes, to use the same fSNPs in ech treatment as the ones used  when the treatments were run using independent models select  NULL}
}
\value{
Saves the summary table in "out" dir as /out/prefix.main.txt. When using tags, saves /out/prefix.tags.lookup.txt. Saves a table of rsnps that couldn't be run.
}
\description{
Run baseqtl for 2 treatments  in one model from different smples
This function allows you to run baseqtl for one gene and multiple pre-selected snps.
}
