% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_GT_paired.R
\name{btrecase.gt.paired.in}
\alias{btrecase.gt.paired.in}
\title{Make inputs for running  baseqtl with known rsnp GT and paired design, optional refbias correction}
\usage{
btrecase.gt.paired.in(
  gene,
  chr,
  snps = 5 * 10^5,
  counts.f,
  covariates = 1,
  additional_cov = NULL,
  e.snps,
  u.esnps = NULL,
  gene.coord,
  vcf,
  le.file,
  h.file,
  population = c("EUR", "AFR", "AMR", "EAS", "SAS", "ALL"),
  nhets = 5,
  min.ase = 5,
  min.ase.het = 5,
  tag.threshold = 0.9,
  out = ".",
  prefix = NULL,
  model = c("both", "NB-ASE", "NB"),
  prob = NULL,
  prior = NULL,
  ex.fsnp = NULL,
  AI_estimate = NULL,
  pretotalReads = 100
)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{chr}{chromosome where the gene is, example chr=22}

\item{snps}{either cis-window or character vector with pos:ref:alt allele for each snp, defaults to cis-window}

\item{counts.f}{path to files with filtered counts: rows genes, first col gene_id followed by samples, prepared in inputs.R per treatment}

\item{covariates}{path to matrix of covariates prepared in inputs.R, if no covariates, covariates =1, default}

\item{additional_cov}{full name to file with first column sample names and additional columns gene independent covariates, defaults to NULL}

\item{e.snps}{path to file listing exonic snps for the chromosome where the gene is, prepared in input.R}

\item{u.esnps}{whether to use unique exonic snps per gene, defaults to NULL when it is not necessary if strand info is known}

\item{gene.coord}{path to file listing gene coordinates and exons, prepared in input.R}

\item{vcf}{path to vcf files with ASE and GT for the chromosome where the gene is, for each treatment, same order as counts.f}

\item{le.file}{path to gz legend file (legend.gz) for the chromosome of interest for the reference panel (snp description)}

\item{h.file}{path to gz haplotpe file for the chromosome of interest for the reference panel (haplotypes for all samples in reference panel)}

\item{population}{ethnicity to get EAF for rsnp: AFR AMR EAS EUR SAS ALL, defaults to EUR}

\item{nhets}{minimun number of het individuals in order to run the minumn model (NB only), defaults to NULL}

\item{min.ase}{minimun number of ASE counts for an individual in order to be included, defaults to 5}

\item{min.ase.het}{minimun number of het individuals with the minimun of ASE counts in order to run the ASE side of the model, defaults to NULL}

\item{tag.threshold}{numeric with r2 threshold (0-1) for grouping snps to reduce the number of running tests, to disable use "no"}

\item{out}{path to save outputs, default to current directory}

\item{prefix}{optional prefix for saving tables, if NULL gene_id.eqtl will be used}

\item{model}{whether to run NB-ASE (full model negative binomial and allele specific counts),NB (negative binomial only) or both (NB-ASE and NB for those associations with no ASE information)}

\item{prob}{number p∈(0,1) indicating the desired probability mass to include in the intervals, defaults to 0.99 and 0.95 quantiles}

\item{prior}{named list: mean= vector with the mean of Gaussians, sd= vector with Gaussians sd for eQTL effect prior, mix=vector with mixing proportions. Defaults to NULL, mixture of 3 components with mean (0,0,0) ;sd  c( 0.0436992, 0.3492696, 0.4920049); and mixing proportions  c(0.955,2*0.015, 0.015).}

\item{ex.fsnp, }{if character: vector with pos:ref:alt for fsnps to exclude,  defaults to NULL}

\item{AI_estimate}{full name to data table with AI estimates for reference panel bias for fSNPs, defaults to NULL}

\item{pretotalReads}{numeric indicating a cut-off for total initial reads to consider AI estimates, defaults to 100}

\item{min.ase.n}{minimun number individuals with the minimun of ASE counts, defaults to NULL}
}
\value{
data.table with summary of gene-snp associations. Saves the summary table in "out" dir as /out/prefix.main.txt. When using tags, saves /out/prefix.tags.lookup.txt. Saves a table of excluded rsnps.
btrecase.gt.paired.in()
}
\description{
This function allows you to run baseqtl for one gene and multiple pre-selected snps. When there is no enough information to ASE counts or rSNP is not in the reference panel, the function will run bayesian negative binomial model only by default.
}
