% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{cl_bcfq}
\alias{cl_bcfq}
\title{command line to extract any field via bcftools query, allow to select samples or region if required}
\usage{
cl_bcfq(
  vcf,
  chr = NULL,
  st = NULL,
  end = NULL,
  samples = NULL,
  f.arg = NULL,
  part = c("body", "header")
)
}
\arguments{
\item{vcf}{full path to vcf or bcf, compressed or uncompressed}

\item{chr}{chromosome to extract}

\item{st}{start position to extract}

\item{end}{end position to extract}

\item{samples}{character vector with samples to extract, defaults=NULL}

\item{f.arg}{character vector with -f argument for bcftools query, defaults to GT and ASE}

\item{part}{whether to extract body or header of vcf}
}
\value{
character vector with bcftools command
cl_bcfq()
}
\description{
This function allows you to write the command to extract any field for a region of a vcf, allows sample selection
}
\keyword{ASE}
\keyword{GT}
\keyword{bcftools}
\keyword{command}
\keyword{line}
\keyword{query}
