% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{cl_coord}
\alias{cl_coord}
\title{get start and end of cis-window per gene to make a bcftools query from a vcf file}
\usage{
cl_coord(file, chr, chrCol = 2, gene, stCol = 4, endCol = 5, cw = 5e+05)
}
\arguments{
\item{file}{full path to file with gene coordinates, as prepared in inputs.R}

\item{chr}{chromosome to extract}

\item{gene}{gene id}

\item{cw}{length of cis-window, defaults to 5*10^5}
}
\value{
vector with start and end of cis-window for the selected gene
cl_coord()
}
\description{
This function allows you to write the command to extract GT and ASE for a region of a vcf
}
\keyword{bcftools}
\keyword{cis}
\keyword{command}
\keyword{line}
\keyword{query}
\keyword{window}
