% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{cor2}
\alias{cor2}
\title{Get correlation of haps from reference panel, to input in tag from GUESSFM, from Chris}
\usage{
cor2(x)
}
\arguments{
\item{x}{matrix (example haplotypes for the snps, cols snps, rows samples)}
}
\value{
matrix of correlations
cor2
}
\description{
calculates correlation by cols
}
\keyword{correlation}
\keyword{haplotypes}
\keyword{panel}
\keyword{reference}
