% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{filt.fsnp}
\alias{filt.fsnp}
\title{function for filtering fsnps, allows to select filtering if n=0 or n=m (likely GT error), also option to remove fsnps with less than cut-off total ASE counts in all samples, INPUT matrix}
\usage{
filt.fsnp(c.ase, ase = 5, min.ase.snp = 5, n = 5, gt.err = NULL, rem = "yes")
}
\arguments{
\item{c.ase}{matrix output of tot.ase_counts}

\item{ase}{total ase cut-off default is 5 counts, trecase default 5}

\item{min.ase.snp}{per snp ase cut-off}

\item{n}{minimun number of individuals het for fsnp with counts >=ase, trecase default 5}

\item{gt.err}{optional parameter, if "yes" it sets to m=0 when corresponding n=0 or n=m (likely GT error, homo called het), defaults to NULL}

\item{rem}{whether to exclude fsnps with less than min.ase.snps ASE counts in all samples, defaults to "yes", otherwise set it to null}
}
\value{
matrix with rownames sample name, cols: n and m counts for each fSNP, option to remove snps with ASE counts below cut-off in all individuals. Also ASE count below threshold (ase.min.snp) are replaced with 0, so will be excluded in subsequent analysis

filt.fsnp()
}
\description{
This function allows you to test whether a fsnp has enough het inds with a certain number of ASE counts
}
\keyword{counts}
\keyword{filtering}
\keyword{m}
\keyword{matrix}
