% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{filt.rsnp}
\alias{filt.rsnp}
\title{function for filtering input for stan, also option to remove fsnps with 0 m counts in all samples}
\usage{
filt.rsnp(geno.exp, ase = 5, n = 5, rem = NULL)
}
\arguments{
\item{geno.exp}{data table for an element of output list tot.ase_counts}

\item{ase}{ase cut-off default is 5 counts, trecase default 5}

\item{n}{minimun number of individuals het for rsnp with counts >=ase, trease default 5}

\item{rem}{whether to exclude fsnps with 0 m counts in all samples, defaults to NULL}
}
\value{
data table with sample name, GT rsnp, total counts, n and m counts for each fSNP, same data table outputted as an element of tot.ase_counts

filt.rsnp()
}
\description{
This function allows you to test whether a rsnp has enough het inds with a certain number of ASE counts
}
\keyword{filtering}
\keyword{input}
\keyword{stan}
