% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{fsnp.prep}
\alias{fsnp.prep}
\title{helper function for formatting input for baseqtl stan no GT and missing values, prepares inputs only dependent on fsnps to feed stan.trecase.rna.noGT.eff}
\usage{
fsnp.prep(rp.f, f.ase, c.ase, min.ase = 5, min.ase.n = 5)
}
\arguments{
\item{rp.f}{matrix of haplotypes from reference panel for fsnps, rows=snps, cols samples from reference panel}

\item{f.ase}{data.table with GT and ASE for fSNPs, output from gt.as, same fsnps in same order as rp.f}

\item{c.ase}{output from tot.ase for the rsnp id}

\item{min.ase}{ase cut-off default is 5 counts, trecase default 5}

\item{min.ase.n}{minimun number of individuals with suffiicent ase counts}
}
\value{
list of 1) mis.sample, logical matrix, for each sample indicates which fnsps are missing (TRUE) 2) u.f, matrix with unique cols from mis.samples, 3) f.comb list to relate u.f with name of snps, each element is a vector with name colname in u.f, and values names of fsnps that are not missing.

fsnp.prep()
}
\description{
This function allows you to format input for stan related to fsnps only
}
\keyword{input}
\keyword{stan}
