% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{fsnp.prep2}
\alias{fsnp.prep2}
\title{Function for formatting input for baseqtl no GT and missing values, prepares inputs only dependent on fsnps to feed stan.trecase.rna.noGT.eff}
\usage{
fsnp.prep2(rp.f, f.ase, c.ase, min.ase = 5, min.ase.n = 5, ai = NULL)
}
\arguments{
\item{rp.f}{matrix of haplotypes from reference panel for fsnps, rows=snps, cols samples from reference panel}

\item{f.ase}{data.table with GT and ASE for fSNPs, output from gt.as, same fsnps in same order as rp.f}

\item{c.ase}{output from tot.ase for the rsnp id}

\item{min.ase}{ase cut-off default is 5 counts, trecase default 5}

\item{min.ase.n}{minimun number of individuals with suffiicent ase counts}

\item{ai}{data table with allelic imbalance estimate for reference panel bias correction, defaults to NULL}
}
\value{
list of 1) m= numeric with total ase counts. 2) n= list of matrices: each matrix has n.counts, rows samples, cols haplotypes compatible with the reference panel. 3) NB= list of data tables, each data table summarizing samples with the haplotypes compatible with RP, to make NB input, 4) f.comb, list relating names on n list and NB list to fsnps used for making haplotypes

fsnp.prep2()
}
\description{
This function allows you to format input for stan related to fsnps only
}
