% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_GT_paired.R
\name{g.for}
\alias{g.for}
\title{Aux for returning genotype info}
\usage{
g.for(dt)
}
\arguments{
\item{dt}{data table with inputs}
}
\value{
vector to return to stan
g.for()
}
\description{
Aux for returning genotype info
}
