% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{hap.p.no.gt}
\alias{hap.p.no.gt}
\title{sub-function for formatting input for stan with no GT for rsnp}
\usage{
hap.p.no.gt(M, y)
}
\arguments{
\item{M}{sparseMatrix  with hap pairs and GT prob for reference panel}

\item{y}{character vector with haplotype pairs (separeated by ",") to match in x (allowing swaps)}
}
\value{
character vector with names y and value the rownumber in M, if no match row=0

hap.p.no.gt()
}
\description{
This function allows you to check whether hap.pairs (rownames of matrix M)  are in a input vector allowing for swap
}
\keyword{genotype}
\keyword{pH}
\keyword{rsnp}
\keyword{stan}
\keyword{unknown}
