% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{hap.p.no.gt2}
\alias{hap.p.no.gt2}
\title{sub-function for formatting input for stan with no GT for rsnp, similar to hap.p.no.gt but avoids re-doing fisrt step (s.M)}
\usage{
hap.p.no.gt2(s.M, M, y)
}
\arguments{
\item{s.M}{named vector with names rownames(M) and value swapped rownames(M)}

\item{M}{matrix  with hap pairs and GT prob for reference panel}

\item{y}{character vector with haplotype pairs (separeated by ",") to match in x (allowing swaps)}
}
\value{
character vector with names y and value the rownumber in M, if no match row=0

hap.p.no.gt2()
}
\description{
This function allows you to check whether hap.pairs (rownames of matrix M)  are in a input vector allowing for swap
}
\keyword{genotype}
\keyword{pH}
\keyword{rsnp}
\keyword{stan}
\keyword{unknown}
