% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various_cin.R
\name{hap_sam2}
\alias{hap_sam2}
\title{Get haps of fsnp plus rSNP, improved for more consistent output}
\usage{
hap_sam2(x, w = NULL, y = NULL, z = NULL)
}
\arguments{
\item{x}{data table of fsnps, rows fsnps and columns phased GT, plus any additional}

\item{w}{sample names, default to NULL as can be deduced from x}

\item{y}{rsnp id, id="POS:REF:ALT"}

\item{z}{data table or rsnps, rows fsnps and columns phased GT, plus any additional}
}
\value{
list with 2 matrices, first for hap1: row samples and each col GT for fsnps and last rsnp
hap_sam2
}
\description{
extract haps from vcf file for fsnps and 1 rsnp
}
\keyword{haps}
\keyword{sample}
