% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{haps.range}
\alias{haps.range}
\title{Extract haps from hap file for the whole cis-window (range)}
\usage{
haps.range(file1, file2, cw, population = "EUR", maf = 0.05)
}
\arguments{
\item{file1}{full path to file to legend.gz file}

\item{file2}{full path to hap.gz file}

\item{cw}{vector start and end position of snps to extract within the range}

\item{population}{ethnicity to set a cut-off for maf: AFR AMR EAS EUR SAS ALL}

\item{maf}{cut-off for maf}
}
\value{
matrix with haplotypes as in hap format (each col one hap), rownames are snp id
}
\description{
This function allows you extract haplotypes for a range of snps
}
\keyword{haplotypes}
\keyword{panel}
\keyword{range}
\keyword{reference}
\keyword{snps}
