% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{haps}
\alias{haps}
\title{Generating all possible haplotypes for a set of sample and their corresponding AS}
\usage{
haps(x, y)
}
\arguments{
\item{x}{data table with GT and AS for fSNPs and rSNPs}

\item{y}{name of samples to make haps}
}
\value{
list
haps()
}
\description{
This function allows you to compute all possible haplotypes for an individual for a set of fSNPs and each rSNP
}
