% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{haps_sub}
\alias{haps_sub}
\title{Generating all possible haplotypes for a sample: sub function for haps: based on Chris F code and on binomial theorem: (a+b)^n= Sumk=0,n nchoosek a^k.b^(n-k)}
\usage{
haps_sub(z, het.rows)
}
\arguments{
\item{z}{data table with GT for 1 individual}

\item{het.rows}{positions in z that correspond to het SNPs}
}
\value{
list of lists. Inner list haplotypes per individual, outer list each individual
haps_sub()
}
\description{
This function allows you to compute all possible haplotypes for an individual for a set of SNPs
}
\keyword{all}
\keyword{haplotype}
