% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_GT.R
\name{help.tags}
\alias{help.tags}
\title{help with tags}
\usage{
help.tags(gene, tag.threshold, rs, prefix, rec.rs, out)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{tag.threshold}{numeric with r2 threshold (0-1) for grouping snps to reduce the number of running tests, to disable use "no"}

\item{rs}{data table with genotyps}

\item{prefix}{optional prefix for saving tables, if NULL gene_id.eqtl will be used}

\item{rec.rs}{data table with recoded genotypes}

\item{out}{path to save outputs, default to current directory}
}
\value{
saves tags and tagged SNPS and return tags id
}
\description{
help with tags
}
