% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{in.neg.beta.noGT.eff2}
\alias{in.neg.beta.noGT.eff2}
\title{prepare input for stan neg.beta.noGT.rsnp.priors.eff2.stan}
\usage{
in.neg.beta.noGT.eff2(x, covar = 1)
}
\arguments{
\item{x}{output list from  stan.trecase.rna.noGT.eff2}

\item{covar}{matrix or numeric with covariates, rows individuals, cols covariates. Individuals in same order as in x, do not include col of ones for intercept. Defaults to intercept only}
}
\value{
list to input to neg.beta.noGT.rsnp.priors.eff.stan
in.neg.beta.noGT.eff2()
}
\description{
This function allows you to prepare inputs for stan neg.beta.noGT.rsnp.priors.eff2.stan
}
