% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{in.neg.beta.prob.eff2}
\alias{in.neg.beta.prob.eff2}
\title{prepare input for stan neg.beta.prob.priors.eff2.stan}
\usage{
in.neg.beta.prob.eff2(x, covar = 1)
}
\arguments{
\item{x}{element from output list from stan.neg.beta.prob.eff}

\item{covar}{matrix or numeric with covariates, rows individuals, cols covariates. Same order as in x, do not include col of ones for intercept. Defaults to intercept only}
}
\value{
list to input to stan neg.beta.ase.prob.priors.eff2.stan
in.neg.beta.prob.eff2()
}
\description{
This function allows you to prepare inputs for neg.beta.prob.phasing.priors.eff.test.stan
}
