% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{input.neg.only.bj}
\alias{input.neg.only.bj}
\title{prepare input for stan neg.only}
\usage{
input.neg.only.bj(DT1, DT2, covar = 1)
}
\arguments{
\item{DT1}{data table with total counts, 1 row for gene, only cols for counts, as in top.rsnp.22.r}

\item{DT2}{data table with rsnp GT (row) for samples (colums), extra cols allowed}

\item{covar}{data table with covariate information, exclude intercept}
}
\value{
list to input to neg.only.bj.stan
input.neg.only.bj()
}
\description{
This function allows you to prepare inputs for neg.only
}
\keyword{input}
\keyword{neg}
\keyword{only}
\keyword{stan}
