% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_noGT_2T.R
\name{list.err}
\alias{list.err}
\title{convenient wrap for testing function per skin}
\usage{
list.err(l, txt)
}
\arguments{
\item{l}{list to test, list has dta tables and I am testing for rows>0 in both elements}

\item{txt}{character with text for error message}
}
\value{
character with error message
list.err()
}
\description{
convenient wrap for testing function per skin
}
