% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{list.help}
\alias{list.help}
\title{helper function to prepare list of inputs for stan.trecase.eff2}
\usage{
list.help(g, gt.r, n.mat, DT, m, ai.mat = NULL, vai.mat = NULL)
}
\arguments{
\item{g}{genotype rsnp in scale 0,1,-1,2}

\item{gt.r}{named numeric vector genotype for rsnp in scale 0,1,-1,2 for all samples, names are sample ID}

\item{n.mat}{matrix of n counts for each hap pair of fsnps concordant with reference panel (cols) and rows sample id}

\item{DT}{data table with hap pairs of fsnp and rsnp by genotype rsnp, and their respective probabilibities, as made in stan.trecase.eff2}

\item{m}{character vector with total ASE counts, names sample ID}

\item{ai.mat}{matrix of allelic imbalance estimates as logit for each hap pair of fsnps concordant with reference panel (cols) and rows sample id, defaults to NULL}

\item{vai.mat}{matrix of variance for allelic imbalance estimates for each hap pair of fsnps concordant with reference panel (cols) and rows sample id, defaults to NULL}
}
\value{
list of list of 1) list of n counts, named by sample id, 2) list of p(H|G) named by sample id 3) optional allelic imbalance named by sample id

list.help()
}
\description{
This function allows you to format selected n counts and  pH|G based on GT
}
\keyword{GT}
\keyword{input}
\keyword{rsnp}
\keyword{stan}
