% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{mat.col}
\alias{mat.col}
\title{Function for making matrix of P(h|GT fSNPs) (colSums = 1) starting from matrix of h vs GT (fsnps+rsnp)}
\usage{
mat.col(M)
}
\arguments{
\item{M}{matrix from Matrix pack}
}
\value{
sparseMatrix from Matrix pack
mat.col()
}
\description{
collapsing matrix by columns
}
