% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{mat.cond}
\alias{mat.cond}
\title{Function for making matrix of P(H snps|GT snps) (colSums = 1) starting from matrix p(H)}
\usage{
mat.cond(M)
}
\arguments{
\item{M}{matrix from Matrix package}
}
\value{
matrix from Matrix package
mat.cond()
}
\description{
conditional matrix by cols, divides each col of M by its sum
}
\keyword{conditional}
\keyword{matrix}
