% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various_cin.R
\name{name}
\alias{name}
\title{Format txt converted vcf files: one file per chr}
\usage{
name(file, head = NULL)
}
\arguments{
\item{file}{file name with full path to vcf}

\item{head}{path and file with header, defaults to prefix.header.txt}
}
\value{
named list with info from vcf files
name()
}
\description{
This function allows you to get good format from vcf text.
}
\keyword{format}
\keyword{vcf}
