% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{p.hap.pair.dt}
\alias{p.hap.pair.dt}
\title{Sub function for simulating haplotypes with ASE, fSNPs plus one rSNP: calculate probability of population haplotype pairs}
\usage{
p.hap.pair.dt(h)
}
\arguments{
\item{h}{population haplotypes}
}
\value{
data.table with probabilities of haplotype pairs and genotype
p.hap.pair.dt()
}
\description{
This function allows you to calculate frequency of haplotype pairs in population
}
\keyword{haplotype}
\keyword{pairs}
\keyword{probability}
\keyword{simulation}
