% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{p.hap.pair.s}
\alias{p.hap.pair.s}
\title{Calculate probability of population haplotype pairs and return sparse matrix}
\usage{
p.hap.pair.s(h)
}
\arguments{
\item{h}{matrix with population haplotypes}
}
\value{
sparse matrix with probabilities of haplotype pairs and genotype
p.hap.pair.s()
}
\description{
This function allows you to  calculate frequency of haplotype pairs in population, returns a sparse matrix
}
\keyword{haplotype}
\keyword{pairs}
\keyword{probability}
\keyword{simulation}
