% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{prop_het}
\alias{prop_het}
\title{Test of proportions for fsnps}
\usage{
prop_het(f.ase, rp.f, gene)
}
\arguments{
\item{f.ase}{data table with fsnps and GT for samples (f.ase), cols for GT must end with "_GT"}

\item{rp.f}{with rownames fsnps and col haplotypes of fsnps}

\item{gene}{character with gene_id}
}
\value{
matrix with rownames fsnps and colnames OR (odds ratio) and pvalue for Fisher's exact test
}
\description{
This function tests equality of the proportion of hets for fsnps in 2 populations, sample and reference panel
}
\keyword{hets}
\keyword{proportion}
\keyword{test}
