% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{rec.guess}
\alias{rec.guess}
\title{Recode GT from (0,1,-1,2) scale  to GUESSFM scale 0 M, 1 hom ref, 2 het, 3 hom alt}
\usage{
rec.guess(DT)
}
\arguments{
\item{DT}{data table GT coded in trecase scale, rows SNPS, cols samples plus additionals}
}
\value{
matrix with rows samples and cols SNPS
}
\description{
This function allows you to recode GT for input in tag function from GUESSFM
}
\keyword{GUESSFM}
\keyword{recode}
