% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various_cin.R
\name{rec_mytrecase_rSNPs}
\alias{rec_mytrecase_rSNPs}
\title{Selects SNPs and recode GT to my implementation of trecase scale: 0 AA, 1 AB, 2 BB BA -1}
\usage{
rec_mytrecase_rSNPs(x, y, z = NULL, recode = "yes")
}
\arguments{
\item{x}{vector with SNP positions for a chr}

\item{y}{data.table with phased GP (0|1) for each sample, for the chr selected in x input}

\item{z}{vector with order fo samples, defaults to NULL}

\item{recode}{whether not to recode, defaults to yes}
}
\value{
data table with recoded GT per sample
rec_mytrecase_rSNPs
}
\description{
recodes fixed genotypes to my implementation of trecase for selected SNPs, samples with "_GT" suffix.
}
\keyword{GT}
\keyword{m=trecase}
\keyword{rSNP}
\keyword{recode}
