% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{relab}
\alias{relab}
\title{Change names of bp, bn to more general case}
\usage{
relab(a, b, x)
}
\arguments{
\item{a}{suffix for old names}

\item{b}{suffix for new names}

\item{x}{data table to change names}
}
\value{
data table with new names
}
\description{
Make data tble summary more general by replacing bp and bn for bt1 and bt2 when dealing with treatments/conditions
}
