% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{run.stan}
\alias{run.stan}
\title{Run stan model}
\usage{
run.stan(mod, data, pars, probs)
}
\arguments{
\item{mod}{compiled stan model}

\item{data}{input data to run mod}

\item{pars}{character with the name of parameters to extract}

\item{probs}{quantiles to report eQTl effect}
}
\value{
matrix with stan summary
run.stan()
}
\description{
This functions run stan and extract posterior summary for specified quantiles for reporting
}
