% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{snp.eaf}
\alias{snp.eaf}
\title{Extracts snp eaf from legend file for a set of snps}
\usage{
snp.eaf(file1, snps, population = "EUR")
}
\arguments{
\item{file1}{full path to file to legend.gz file}

\item{snps}{character vector with id of snps to extract information, id=POS:REF:ALT}

\item{population}{ethnicity to set a cut-off for maf: AFR AMR EAS EUR SAS ALL}
}
\value{
DT with snp id and EAF from legend file for selected population
}
\description{
This function allows you extract snp info from legend file for a range of snps
}
\keyword{eaf}
\keyword{panel}
\keyword{reference}
\keyword{snp}
