% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{stan.2T}
\alias{stan.2T}
\title{Format a named list of stan output for baseqtl2T.nogt}
\usage{
stan.2T(
  x,
  rtag = NULL,
  gene,
  EAF = NULL,
  info = NULL,
  nfsnps = NULL,
  min.pval = NULL,
  probs = NULL
)
}
\arguments{
\item{x}{list of summaries}

\item{rtag}{optional argument,whether snps were grouped using tag function, defaults to yes}

\item{gene}{gene id of gene under study}

\item{EAF, }{data table with snp and eaf for tag snps, output from snp.eaf (real.data.R), defaults to NULL}

\item{info, }{named vector with info type score for tag snp, names snp_id, defaults to NULL}

\item{nfsnps, }{number of fsnps in analysis, defaults to NULL}

\item{min.pval}{data table with cols fsnp_id, OR (odds ratio), pvalue, gene_id, output from porp.het, for each gene selects the min(pvalue) and add it to output}

\item{probs}{extremes for the posterior probability mass, defaults to 2.5 and 97.5}
}
\value{
DT with formated data
}
\description{
This function allows you to extract the parameter information from a list of stan summaries (output from stan.many.sim)
}
\keyword{2}
\keyword{multiple}
\keyword{snps}
\keyword{stan}
\keyword{tissues}
