% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{stan.ase}
\alias{stan.ase}
\title{sub-function for formatting input for baseqtl no GT, deals with genotype of rsnp and p(Grsnp|G fsnp) & p(H|G) to make stan input for ASE}
\usage{
stan.ase(M, M.cond, s.M, n.mat, ai0.mat = NULL, vai0.mat = NULL)
}
\arguments{
\item{M}{matrix of haplotypes from reference panel for fsnps and rsnp, rows=hap.pairs, cols GT}

\item{M.cond}{matrix haplotypes from reference panel, rows as M, cols GT, colsums=1}

\item{s.M}{named vector with names rownames(M) and value swapped rownames(M), to identify swaps, done in stan.rsnp.noGT.eff}

\item{n.mat}{matrix with hap for fsnps and ase counts, output from fsnp.prep or stan.fsnp.noGT}

\item{ai0.mat}{matrix with estimates for allelic imbnalance for reference bias correction, defaults to NULL
@param vai0.mat matrix with variance estimates for allelic imbnalance for reference bias correction, defaults to NULL}
}
\value{
list of stan input: list of 1) list with vectors of n.counts, 2) list of vectors of  p(H(f+r)|Gf) , 3) list of vectors of possible genotype rsnp in scale 0,1,-1,2, each vector one sample, 4)optional list with AI reference bias estimates
stan.ase()
}
\description{
This function allows you to format input for stan ASE per individual: adapted from stan.rsnp.noGT.eff
}
\keyword{input}
\keyword{rsnp}
\keyword{stan}
