% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{stan.cov}
\alias{stan.cov}
\title{sub-function to deal with covariates when preparing input for stan}
\usage{
stan.cov(N, covar = 1)
}
\arguments{
\item{N, }{number of individuals with covariate information}

\item{covar}{matrix or numeric with covariates, rows individuals, cols covariates. Same order as in x, do not include col of ones for intercept. Defaults to intercept only}
}
\value{
matrix with covariates compatible with stan: two cols with ones plus covs, first col will be ignored by stan but allows to enter data as matrix avoidind vector when no covs
stan.cov()
}
\description{
This function allows you to deal with covariates for stan input
}
