% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{stan.fsnp.noGT.eff}
\alias{stan.fsnp.noGT.eff}
\title{Function for formatting input for baseqtl stan no GT, deals with haps, n counts and m counts of fSNPs to streamline computation, optional to include allelic imbalance estimate for dealing with reference panel bias}
\usage{
stan.fsnp.noGT.eff(
  rp.f,
  f.ase,
  c.ase,
  NB = "yes",
  min.ase = 5,
  min.ase.n = 5,
  ai = NULL
)
}
\arguments{
\item{rp.f}{matrix of haplotypes from reference panel for fsnps, rows=snps, cols samples from reference panel}

\item{f.ase}{data.table with GT and ASE for fSNPs, output from gt.as, same fsnps in same order as rp.f}

\item{c.ase}{output from tot.ase for the rsnp id}

\item{NB, }{whether to return NB element of output, defalts to "yes" as it is required for noGT, use "no" for GT input preparation}

\item{min.ase}{ase cut-off default is 5 counts, trecase default 5}

\item{min.ase.n}{minimun number of individuals with sufficient ase counts, trease default 5}

\item{ai}{data table with AI estimates per SNP, defaults to NULL}
}
\value{
list of 1) m= numeric with total ase counts. 2) n= list of matrices: each matrix has n.counts, rows samples, cols haplotypes compatible with the reference panel. 3) NB= data table summarizing samples with the haplotypes compatible with RP, to make NB input, 

stan.fsnp.noGT.eff()
}
\description{
This function allows you to format input for stan per individual: precomputes m counts and n counts for each individual based on haplotypes compatible with genotypes of the fsnps based on the reference panel
}
\keyword{fsnp}
\keyword{input}
\keyword{stan}
