% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{stan.nb}
\alias{stan.nb}
\title{sub-function for formatting input for baseqtl stan no GT, deals with genotype of rsnp and p(Grsnp|G fsnp) & p(H|G) to make stan input for negative binomial}
\usage{
stan.nb(counts.g, M, M.cond, NB)
}
\arguments{
\item{counts.g}{DT of total gene counts, colnames sample names}

\item{M}{matrix of haplotypes from reference panel for fsnps and rsnp, rows=hap.pairs, cols GT}

\item{M.cond}{matrix haplotypes from reference panel, rows as M, cols GT, colsums=1}

\item{NB}{data table with summarize information for neg binom side, output from fsnp.prep or stan.fsnp.noGT}
}
\value{
list of stan input: list of 1) counts: data table with total counts  and 2) list p.g with p(genotype) rsnp per sample.
stan.nb()
}
\description{
This function allows you to format input for stan Nb per individual: adapted from stan.rsnp.noGT.eff
}
\keyword{input}
\keyword{rsnp}
\keyword{stan}
