% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{sum.g}
\alias{sum.g}
\title{helper function to summarise genotypes of fnsps across individuals}
\usage{
\method{sum}{g}(h.short)
}
\arguments{
\item{h.short}{list with 2 elements, first character named vector with haplotype 1 for fsnps, if 3 fsps each element will look as "010", etc. names are sample ID. Second element corresponds to hap2}
}
\value{
data table with columns: hap1, hap2, g.fsnps (GT fsnp), ind: row number from h.short for the individuals that have that hap combination, hap1.2, hap1 and hap2 separated by ",", row.M row number in which that hap combination is found in input M

sum.g()
}
\description{
This function allows you to format summarise the genotypes of fnsps across individuals
}
\keyword{fsnps}
\keyword{summary}
