% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{sw.no.gt}
\alias{sw.no.gt}
\title{sub-function for formatting input for stan with no GT for rsnp}
\usage{
sw.no.gt(M, M.cond, DT, m.trim, i, j, n.col, m.col)
}
\arguments{
\item{M}{matrix with p(H) row hap.pairs and cols GT for fsnps plus rsnp from reference panel}

\item{M.cond}{matrix with p(H|GT fsnps), row hap.pairs and cols GT for fsnps from reference panel, output from mat.col}

\item{DT}{data table with cols: GT.ob (observed GT,h1.2, hap1,hap2 and rows.comp row in M that matches h1.2, constructed in stan.ase.no.gt}

\item{m.trim}{matrix with ASE information}

\item{i}{index for selecting rows of m.trim}

\item{j}{GT.ob, to work within loop}

\item{n.col}{names for cols in mat with n counts}

\item{m.col}{names for cols in mat with m counts}
}
\value{
list with partial inputs for stan p= collection of p(H), n=n counts, r= gt for rsnp in scale 0,1,-1,2

sw.no.gt()
}
\description{
This function allows you to get p(H), n and GT for rsnp for non-observed hap of fsnps and rsnp compatibles with observed GT for fsnps and haplotypes in reference panel
}
\keyword{genotype}
\keyword{input}
\keyword{rsnp}
\keyword{stan}
\keyword{unknown}
