% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{tag.noGT}
\alias{tag.noGT}
\title{Derive tag SNPs for a SnpMatrix object using heirarchical clustering}
\usage{
tag.noGT(X, tag.threshold = 0.99, quiet = FALSE, method = "single")
}
\arguments{
\item{X}{matrix of haplotypes, each rows is an observed hap, cols snps}

\item{tag.threshold}{threshold to cut tree, default=0.99}

\item{quiet}{if FALSE (default), show progress messages}

\item{method}{method used for heirarchical clustering.  See hclust for options.}
}
\value{
character vector, names are \code{snps}, values are the tag for each SNP
}
\description{
Uses complete linkage and the \code{\link{hclust}} function to define clusters,
then cuts the tree at 1-tag.threshold. Based on Chris Wallace's function tags but allowing for haplotype input.
}
