% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_GT_paired.R
\name{treat.form}
\alias{treat.form}
\title{Aux for formatting 2 tretmenat inputs}
\usage{
treat.form(l, dt, n)
}
\arguments{
\item{l}{list with stan input for 2 treatments for 1 snps}

\item{dt}{data table to use for formatting}

\item{n}{name of elements on l to join}
}
\value{
data table with inputs for stan
treat.form()
}
\description{
Aux for formatting 2 tretmenat inputs
}
