% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{u.hap.pairs}
\alias{u.hap.pairs}
\title{Get unique pairs of haplotypes from a set of haplotypes}
\usage{
u.hap.pairs(h)
}
\arguments{
\item{h}{haplotype set}
}
\value{
data table of unique combinations of pairs
u.hap.pairs()
}
\description{
This function allows you to get a unique set of haplotypes
}
