% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{u.hap.pairs2}
\alias{u.hap.pairs2}
\title{Get unique pairs of haplotypes starting from all possible combinations of haplotype pairs}
\usage{
u.hap.pairs2(uhaps)
}
\arguments{
\item{uhaps}{haplotype set with unique entries}
}
\value{
data table of unique combinations of pairs
u.hap.pairs2()
}
\description{
This function allows you to get a unique set of haplotype pairs starting from a unique set of haplotypes
}
