% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.eff.R
\name{unload.ddl}
\alias{unload.ddl}
\title{deal with too many ddls loaded when running stan many times: https://github.com/stan-dev/rstan/issues/448}
\usage{
unload.ddl(model)
}
\arguments{
\item{model}{name of object returned by stan_model, to avoid recompilation.}
}
\value{
unload old ddls
unload.ddl()
}
\description{
This function allows you to unload not in use ddls
}
\keyword{bug}
\keyword{ddls}
\keyword{stan}
