% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{var.e}
\alias{var.e}
\title{wrap functions to calculate var(E(G)), var(G), r2=var(E(G))/var(G) and abs(`log2(aFC)_mean.ngt`-`log2(aFC)_mean.gt`) into a data.frame}
\usage{
var.e(genes = NULL, path, pattern, noGT, le.file, hap.file)
}
\arguments{
\item{genes, }{character vector of genes to select inputs from, defaults to NULL to use all genes run by model}

\item{path}{charcater vector with path to files with stan.input}

\item{pattern}{character vector with pattern to match}

\item{noGT}{data table with Gene_id and tags to extract gene/snps pairs}

\item{le.file}{path to legend file for specific chromosome}

\item{hap.file}{path to haplotype file for specific chromosome}
}
\value{
input data table noGT with new cols: var(E(G)), var(G) and r2
var.e()
}
\description{
This function allows you to to calculate var(E(G)), var(G), r2=var(E(G))/var(G) and abs(`log2(aFC)_mean.ngt`-`log2(aFC)_mean.gt`) into a data.frame
}
\keyword{inputvar(E(G))}
\keyword{stan}
