% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{var.eg}
\alias{var.eg}
\title{var.eg}
\usage{
var.eg(l)
}
\arguments{
\item{l}{list, each element is for one individual, for each ind a vector with p(G)}
}
\value{
character vector, names are \code{snps}, values are the tag for each SNP
}
\description{
Calculate variance of expected genotype from a list with each element an individual and for each individual a vector of probabilities for GT=0,1 or 2.
}
