% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{var.rp}
\alias{var.rp}
\title{Calculates variance of genotype for a group of snps based on reference panel}
\usage{
var.rp(file1, file2, x)
}
\arguments{
\item{file1}{path to legend file}

\item{file2}{path to hap file}

\item{x}{vector with snps coded as pos:ref:alt}
}
\value{
named vector with var(G)
}
\description{
Calculates variance of genotype for a group of snps based on reference panel
}
