% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{vcf.gt.qc}
\alias{vcf.gt.qc}
\title{Check if gt.ase, output from vcf_w has phased GT in GT field, can also save a new vcf with excluding wrong GT by snp or by sample, as required.}
\usage{
vcf.gt.qc(
  gt.ase,
  exclude = c("snps", "samples"),
  vcf.path,
  path = ".",
  vcf.out = "chr22.GTqc"
)
}
\arguments{
\item{gt.ase}{object returned from vcf_w}

\item{exclude}{optional argument,removes entries with wrong GT format by snps or by sample, options "snps" or "samples"}

\item{vcf.path}{path and file name of original vcf, argument used for preparing new vcf with wrong GT entries removed}

\item{path}{optional, path to save new vcf excluding wrongly formatted GT, the default corresponds to the working directory}

\item{vcf.out}{optional, prefix for new vcf with wrong GT entries removed}
}
\value{
named vector when the only argument used is gt.ase. The vector gives the total number of snps, number of snps with wrong GT format, total number of samples and number of samples with wrong GT format. When all arguments are used it saves and indexes a new vcf excluding wrongly GT entries by snps or samples in format vcf.gz. In this mode the function returns a DT with the chr:pos:ref:alt of snps excluded or the name of the samples excluded.
vcf.gt.qc()
}
\description{
This function allows you to test wether a vcf_w returned object is correctly formatted in GT field with the option to list snps or samples with wrong GT in format to be excluded from
}
\keyword{gt}
\keyword{qc}
\keyword{vcf}
