% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{vcf_cl}
\alias{vcf_cl}
\title{Open vcf in R via bcftools query}
\usage{
vcf_cl(cl1, cl2, sep = " ")
}
\arguments{
\item{cl1}{command line bcftools query to extract body of vcf}

\item{cl2}{command line bcftools query to extract header of vcf}

\item{sep}{separation fields based on bcftools query, defaults =" "}
}
\value{
data table
vcf_cl()
}
\description{
This function allows you to extract part of a vcf file and read it as data table
}
\keyword{fread}
\keyword{vcf}
