% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.data.R
\name{vcf_w}
\alias{vcf_w}
\title{Extract GT and ASE for a gene}
\usage{
vcf_w(
  vcf,
  chr = NULL,
  st = NULL,
  end = NULL,
  samples = NULL,
  f.arg = NULL,
  qc = NULL,
  exclude = NULL
)
}
\arguments{
\item{vcf}{full path to vcf}

\item{chr}{chromosome to extract, null for whole vcf}

\item{st}{start position to extract, null for whole vcf}

\item{end}{end position to extract, null for whole vcf}

\item{samples}{character vector with samples to extract, defaults=NULL}

\item{f.arg}{character vector with -f argument for bcftools query, defaults to GT and ASE}

\item{qc}{use function for qc purpose only}

\item{exclude}{whether to return a list with snps excluded from vcf (all homo or missing)}
}
\value{
data table with GT and ASE, unless !is.null(excluded), returns list with first element data table with GT and ASE and second element a data table with exluded snps.
vcf_w()
}
\description{
This function allows you to extract GT and ASE for a gene, wraper for cl_bcfq and vcf_cl plus some formatting, removes samples with missing data or unphased. also removes snps if homo or missing for all samples. Adds id col to ease matching with legend/hap format
}
