% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_baseqtl_noGT.R
\name{aux.in3}
\alias{aux.in3}
\title{Third aux function to prepare inputs to run baseqtl without known rsnp GT: prepare stan inputs wrap}
\usage{
aux.in3(
  gene,
  ai = NULL,
  case,
  rp.f,
  rp.r,
  f.ase,
  counts.g,
  covariates,
  min.ase = 5,
  min.ase.n = 5,
  info = 0.3,
  snps.ex,
  prefix = NULL,
  out = ".",
  save_input = FALSE
)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{ai}{estimates for allelic imbalance, defaults to NULL}

\item{case, }{data table with ASE counts}

\item{rp.f}{reference panel haps for fSNPs}

\item{rp.r}{ref panel haps for rSNPs}

\item{f.ase}{data table with genotypes for fSNPs}

\item{counts.g}{data table with total counts for gene}

\item{covariates}{path to matrix of covariates prepared in inputs.R. If using gc correction (each gene diffrent value), the matrix has rownames= genes and cols=samples plus extra columns if other covariates are added. If only using lib size or gene independent covariates, rows are samples and columns are covariates. If no covariates, covariates =1, default}

\item{min.ase}{minimun number of ASE counts for an individual in order to be included, defaults to 5}

\item{min.ase.n}{minimun number individuals with the minimun of ASE counts, defaults to 5.}

\item{info}{numeric cut-off for var(E(G))/var(G), var(E(G)) is the expected variance for input and var(G) for reference panel, similar to info score in impute2, defaults to 0.3. rsnps with lower info wont be run by stan.}

\item{snps.ex}{data table with rsnps to exclude from running}

\item{prefix}{character with prefix to add for saving files, defaults to NULL}

\item{out}{path to save outputs, default to current directory}

\item{save_input}{whether to save input to stan model for QC purposes, defaults to FALSE to save disk space. Object ending with "noGT.stan.input.rds" is a named list with each element the inputs for a cis-SNP. For each cis-SNP there is a list of 2 elements: "NB" and "ase". "NB" is a list with elements "counts" and "p.g". "Counts" is a data.table with columns sample names and one row corresponding to the gene, values total read counts. "p.g" is a named list with each element a sample. For each sample there is an array with names genotypes (0,1,2) and values the genotype probabilities. For the "ase" list they are for elements: "m" numeric vector with  total ASE counts per sample. "g" list with each element a sample and for each sample the genoptype of the cis SNP coded as 0,1,2 and -1, with -1 indicating that the alternative allele is in haplotype 1. "p" has the same structure as "g" and indicates the probability for each genotype. "n"  is similar to "g" and "p" but contains the mapped reads to haplotype 2. The file ending with "noGT.fsnps.counts.rds is a matrix with rows samples and columns fSNPS. When a fSNPs ends with ".n" correspond to the counts matching the alternative allele and ".m" indicates the total counts matching the SNP.}
}
\value{
list with stan input
aux.in3()
}
\description{
This function allows you to prepare stan inputs final stages
}
