% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseqtl.GT.R
\name{baseqtl.gt}
\alias{baseqtl.gt}
\title{Run BaseQTL with known rsnp GT, optional refbias correction}
\usage{
baseqtl.gt(
  gene,
  chr,
  snps = 5 * 10^5,
  counts.f,
  covariates = 1,
  e.snps,
  u.esnps = NULL,
  gene.coord,
  vcf,
  le.file,
  h.file,
  population = c("EUR", "AFR", "AMR", "EAS", "SAS", "ALL"),
  nhets = 5,
  min.ase = 5,
  min.ase.het = 5,
  tag.threshold = 0.9,
  out = ".",
  prefix = NULL,
  model = c("both", "NB-ASE", "NB"),
  stan.model = NULL,
  stan.negonly = NULL,
  prob = NULL,
  prior = NULL,
  ex.fsnp = NULL,
  AI_estimate = NULL,
  pretotalReads = 100
)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{chr}{chromosome where the gene is, example chr=22}

\item{snps}{either cis-window or character vector with pos:ref:alt allele for each snp, defaults to cis-window}

\item{counts.f}{path to file with filtered counts: rows genes, first col gene_id followed by samples, prepared in inputs.R}

\item{covariates}{path to matrix of covariates prepared in inputs.R, if no covariates, covariates =1, default}

\item{e.snps}{path to file listing exonic snps for the chromosome where the gene is, prepared in input.R}

\item{u.esnps}{whether to use unique exonic snps per gene, defaults to NULL when it is not necessary if strand info is known}

\item{gene.coord}{path to file listing gene coordinates and exons, prepared in input.R}

\item{vcf}{path to vcf file with ASE and GT for the chromosome where the gene is}

\item{le.file}{path to gz legend file (legend.gz) for the chromosome of interest for the reference panel (snp description)}

\item{h.file}{path to gz haplotpe file for the chromosome of interest for the reference panel (haplotypes for all samples in reference panel)}

\item{population}{ethnicity to get EAF for rsnp: AFR AMR EAS EUR SAS ALL, defaults to EUR}

\item{nhets}{minimun number of het individuals in order to run the minumn model (NB only), defaults to NULL}

\item{min.ase}{minimun number of ASE counts for an individual in order to be included, defaults to 5}

\item{min.ase.het}{minimun number of het individuals with the minimun of ASE counts in order to run the ASE side of the model, defaults to NULL}

\item{tag.threshold}{numeric with r2 threshold (0-1) for grouping snps to reduce the number of running tests, to disable use "no"}

\item{out}{path to save outputs, default to current directory}

\item{prefix}{optional prefix for saving tables, if NULL gene_id will be used}

\item{model}{whether to run NB-ASE (full model negative binomial and allele specific counts),NB (negative binomial only) or both (NB-ASE and NB for those associations with no ASE information)}

\item{stan.model}{compiled stanmodel object with stan model, defaults NULL to use  built-in NB-ASE model. When AI_estimate argument (below) is provided the model corrects for reference panel bias, otherwise it doesn't.}

\item{stan.negonly}{compiled stanmodel object with neg only side, deafults to NULL to use built-in model.}

\item{prob}{number p∈(0,1) indicating the desired probability mass to include in the intervals, defaults to 0.99 -0.95 quantiles}

\item{prior}{named list: mean= vector with the mean of Gaussians, sd= vector with Gaussians sd for eQTL effect prior, mix=vector with mixing proportions. Defaults to NULL, mixture of 2 components with mean (0,0); sd  c( 0.0309, 0.3479); and mixing proportions  c(0.97359164, 0.02640836).}

\item{ex.fsnp, }{character vector with pos:ref:alt for fsnps to exclude,  defaults to NULL which corresponds to no exclusions}

\item{AI_estimate}{full name to data table with AI estimates for reference panel bias for fSNPs, defaults to NULL with no correction}

\item{pretotalReads}{numeric indicating a cut-off for total initial reads to consider AI estimates, defaults to 100}

\item{min.ase.n}{minimun number individuals with the minimun of ASE counts, defaults to NULL}
}
\value{
Saves the summary table in "out" dir as /out/prefix.main.txt. When using tags, saves /out/prefix.tags.lookup.txt. Saves a table of excluded rsnps from model.
}
\description{
This function allows you to run BaseQTL for one gene and multiple pre-selected snps. When there is no enough information to ASE counts or rSNP is not in the reference panel, the function will run bayesian negative binomial model only.
}
