% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseqtl_noGT.R
\name{baseqtl.nogt}
\alias{baseqtl.nogt}
\title{Run baseqtl with unknown rsnp GT and missing values for GT fsnps with optional  reference panel bias correction}
\usage{
baseqtl.nogt(
  gene,
  chr,
  snps = 5 * 10^5,
  counts.f,
  covariates = 1,
  additional_cov = NULL,
  e.snps,
  u.esnps = NULL,
  gene.coord,
  vcf,
  sample.file = NULL,
  le.file,
  h.file,
  population = c("EUR", "AFR", "AMR", "EAS", "SAS", "ALL"),
  maf = 0.05,
  min.ase = 5,
  min.ase.snp = 5,
  min.ase.n = 5,
  tag.threshold = 0.9,
  info = 0.3,
  out = ".",
  prefix = NULL,
  model = NULL,
  model.negonly = NULL,
  prob = NULL,
  prior = NULL,
  ex.fsnp = 0.01,
  AI_estimate = NULL,
  pretotalReads = 100
)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{chr}{chromosome where the gene is, example chr=22}

\item{snps}{either cis-window or character vector with pos:ref:alt allele for each snp, defaults to cis-window}

\item{counts.f}{path to file with filtered counts: rows genes, first col gene_id followed by samples, prepared in inputs.R}

\item{covariates}{path to matrix of covariates prepared in inputs.R. If using gc correction (each gene diffrent value), the matrix has rownames= genes and cols=samples plus extra columns if other covariates are added. If only using lib size or gene independent covariates, rows are samples and columns are covariates. If no covariates, covariates =1, default}

\item{additional_cov}{full name to file with first column sample names and additional columns gene independent covariates, defaults to NULL}

\item{e.snps}{path to file listing exonic snps for the chromosome where the gene is, prepared in input.R}

\item{u.esnps}{whether to use unique exonic snps per gene, defaults to NULL when it is not necessary if strand info is known}

\item{gene.coord}{path to file listing gene coordinates and exons, prepared in input.R}

\item{vcf}{path to vcf file with ASE and GT for the exonic snps for the chromosome where the gene is}

\item{sample.file}{sample file for the reference panel (sample description), to be used if ex.fsnp test is required and population is not the whole reference panel}

\item{le.file}{path to gz legend file (legend.gz) for the chromosome of interest for the reference panel (snp description)}

\item{h.file}{path to gz haplotpe file for the chromosome of interest for the reference panel (haplotypes for all samples in reference panel)}

\item{population}{ethnicity to set a cut-off for maf: AFR AMR EAS EUR SAS ALL, and for testing fSNPs to exclude (ex.fsnp below) if applicabloe, defaults to EUR}

\item{maf}{cut-off for maf, defaults to 0.05}

\item{min.ase}{minimun number of ASE counts for an individual in order to be included, defaults to 5}

\item{min.ase.snp}{minum number of ASE counts for a single snp to be considered, for a particular individual}

\item{min.ase.n}{minimun number individuals with the minimun of ASE counts, defaults to 5.}

\item{tag.threshold}{numeric with r2 threshold (0-1) for grouping snps to reduce the number of running tests, to disable use "no"}

\item{info}{numeric cut-off for var(E(G))/var(G), var(E(G)) is the expected variance for input and var(G) for reference panel, similar to info score in impute2, defaults to 0.3. rsnps with lower info wont be run by stan.}

\item{out}{path to save outputs, default to current directory}

\item{prefix}{optional prefix for saving files, if NULL gene_id.eqtl will be used}

\item{model}{compiled stanmodel object with stan model, defaults NULL which uses  built-in NB-ASE model. When AI_estimate is provided the model corrects for reference panel bias, otherwise id doesn't.}

\item{model.negonly}{compailed stanmodel object with neg only side, deafults to NULL which is built-in model.}

\item{prob}{number p∈(0,1) indicating the desired probability mass to include in the intervals, defaults to 0.99 and 0.95 quantiles}

\item{prior}{named list: mean= vector with the mean of Gaussians, sd= vector with Gaussians sd for eQTL effect prior, mix=vector with mixing proportions. Defaults to NULL, mixture of 2 components with mean (0,0) ;sd  c( 0.0309, 0.3479); and mixing proportions  c(0.97359164, 0.02640836).}

\item{ex.fsnp}{if character: vector with pos:ref:alt for fsnps to exclude, if numeric  p-value cut-off for fSNPs to exclude based on fisher test for suspected genotype error based on comparing the proportion of hets in the sample and reference panel,  defaults to 0.01. If NULL no fSNP will be excluded.}

\item{AI_estimate}{full name to data table with AI estimates for reference panel bias for fSNPs, if NULL no reference panel bias correction is applied}

\item{pretotalReads}{numeric indicating a cut-off for total initial reads to consider AI estimates, defaults to 100}
}
\value{
data.table with summary of gene-snp associations. Saves the summary table in "out" dir as /out/prefix.main.txt. When using tags, saves /out/prefix.tags.lookup.txt. Saves a table of excluded rsnps.
}
\description{
This function allows you to run baseqtl for one gene and multiple pre-selected snps.
}
