% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseqtl_noGT_2T.R
\name{baseqtl2T.nogt}
\alias{baseqtl2T.nogt}
\title{Run baseqtl for 2 treatments  in one model from different smples
This function allows you to run baseqtl for one gene and multiple pre-selected snps.}
\usage{
baseqtl2T.nogt(
  gene,
  chr,
  snps = 5 * 10^5,
  counts.f,
  covariates = 1,
  additional_cov = NULL,
  e.snps,
  u.esnps = NULL,
  gene.coord,
  vcf,
  sample.file = NULL,
  le.file,
  h.file,
  population = c("EUR", "AFR", "AMR", "EAS", "SAS", "ALL"),
  maf = 0.05,
  min.ase = 5,
  min.ase.snp = 5,
  min.ase.n = 5,
  tag.threshold = 0.9,
  info = 0.3,
  out = ".",
  prefix = NULL,
  stan.model = NULL,
  prob = NULL,
  prior = NULL,
  ex.fsnp = 0.01,
  AI_estimate = NULL,
  pretotalReads = 100,
  treatments,
  fishjoin = "joint",
  inference.method = c("sampling", "vb"),
  mc.cores = getOption("mc.cores", 1)
)
}
\arguments{
\item{gene}{gene id for the gene to run}

\item{chr}{chromosome where the gene is, example chr=22}

\item{snps}{either cis-window or character vector with pos:ref:alt allele for each snp, defaults to cis-window}

\item{counts.f}{path to file with filtered counts: rows genes, first col gene_id followed by samples, prepared in inputs.R}

\item{covariates}{path to matrix of covariates prepared in inputs.R. If using gc correction (each gene different value), the matrix has rownames= genes and cols=samples plus extra columns if other covariates are added. If only using lib size or gene independent covariates, rows are samples and columns are covariates. If no covariates, covariates =1, default}

\item{additional_cov}{full name to file with first column sample names and additional columns gene independent covariates, defaults to NULL}

\item{e.snps}{path to file listing exonic snps for the chromosome where the gene is, prepared in input.R}

\item{u.esnps}{whether to use unique exonic snps per gene, defaults to NULL when it is not necessary if strand info is known}

\item{gene.coord}{path to file listing gene coordinates and exons, prepared in input.R}

\item{vcf}{path to vcf file with ASE and GT for the chromosome where the gene is}

\item{sample.file}{sample file for the reference panel (sample description), to be used if ex.fsnp test is required and populatios is not the whole reference panel}

\item{le.file}{path to gz legend file (legend.gz) for the chromosome of interest for the reference panel (snp description)}

\item{h.file}{path to gz haplotpe file for the chromosome of interest for the reference panel (haplotypes for all samples in reference panel)}

\item{population}{ethnicity to get EAF for rsnp: AFR AMR EAS EUR SAS ALL, defaults to EUR. It is also used to exclude fSNPs if frequency is different from reference panel when using no GT and selecting ex.fsnps argument, except if population=ALL}

\item{maf}{cut-off for maf, defaults to 0.05}

\item{min.ase}{minimum number of ASE counts for an individual in order to be included, defaults to 5}

\item{min.ase.snp}{minimum number of ASE counts for a single snp to be considered, for a particular individual}

\item{min.ase.n}{minimum number individuals with the minimum of ASE counts, defaults to 5.}

\item{tag.threshold}{numeric with r2 threshold (0-1) for grouping snps to reduce the number of running tests, to disable use "no"}

\item{info}{numeric cut-off for var(E(G))/var(G), var(E(G)) is the expected variance for input and var(G) for reference panel, similar to info score in impute2, defaults to 0.3. rsnps with lower info wont be run by stan.}

\item{out}{path to save outputs, default to current directory}

\item{prefix}{optional prefix for saving tables, if NULL gene_id will be used}

\item{stan.model}{compiled stanmodel object with stan model, defaults NULL to use  built-in NB-ASE model. When AI_estimate argument (below) is provided the model corrects for reference panel bias, otherwise it doesn't.}

\item{prob}{number p∈(0,1) indicating the desired probability mass to include in the intervals, defaults to 0.99 and 0.95 quantiles}

\item{prior}{named list: mean= vector with the mean of Gaussians, sd= vector with Gaussians sd for eQTL effect prior, mix=vector with mixing proportions. Defaults to NULL, mixture of 2 components with mean (0,0); sd  c( 0.0309, 0.3479); and mixing proportions  c(0.97359164, 0.02640836).}

\item{ex.fsnp}{character vector with pos:ref:alt for fsnps to exclude,  defaults to NULL which corresponds to no exclusions}

\item{AI_estimate}{full name to data table with AI estimates for reference panel bias for fSNPs, defaults to NULL}

\item{pretotalReads}{numeric indicating a cut-off for total initial reads to consider AI estimates, defaults to 100}

\item{treatments}{character vector with the 2 treatments (tissues, diseases) to study, sam order as counts.f, covariates and vcf}

\item{fishjoin}{whether to run Fisher test jointly in all samples or by treatment, defaults to jointly but  for QC purposes, to use the same fSNPs in ech treatment as the ones used  when the treatments were run using independent models select  NULL}

\item{inference.method}{The inference method to use with Stan. Options are \code{c("sampling", "vb", "optimizing")}}

\item{mc.cores}{The number of parallel cores to use when performing tests.}
}
\value{
Saves the summary table in "out" dir as /out/prefix.main.txt. When using tags, saves /out/prefix.tags.lookup.txt. Saves a table of rsnps that couldn't be run.
}
\description{
Run baseqtl for 2 treatments  in one model from different smples
This function allows you to run baseqtl for one gene and multiple pre-selected snps.
}
\examples{
## Retrive input files for running baseqtl2T.nogt
counts.f <- c(
  system.file("extdata/input", "counts_Psoriasis_skin.txt", package = "baseqtl", mustWork = TRUE),
  system.file("extdata/input", "counts_normal_skin.txt", package = "baseqtl", mustWork = TRUE)
)
covariates <- c(
  system.file("extdata/input", "Psoriasis_skin_gc_lib_size.rds", package = "baseqtl", mustWork = TRUE),
  system.file("extdata/input", "normal_skin_gc_lib_size.rds", package = "baseqtl", mustWork = TRUE)
)
vcf <- c(
  system.file("extdata/input", "chr10.ASE.Psoriasis_skin.vcf.gz", package = "baseqtl", mustWork = TRUE),
  system.file("extdata/input", "chr10.ASE.normal_skin.vcf.gz", package = "baseqtl", mustWork = TRUE)
)
e.snps <- system.file("extdata/input", "chr10.fSNPS.ENSG00000178372.txt", package = "baseqtl", mustWork = TRUE)
u.snps <- system.file("extdata/input", "chr10.unique.fSNPS.ENSG00000178372.txt", package = "baseqtl", mustWork = TRUE)
gene.coord <- system.file("extdata/input", "ENSG00000178372_data.txt", package = "baseqtl", mustWork = TRUE)
sample.f <- system.file("extdata/input", "1000GP_Phase3.sample", package = "baseqtl", mustWork = TRUE)
le.file <- system.file("extdata/input", "1000GP_Phase3_subset_chr10.legend.gz", package = "baseqtl", mustWork = TRUE)
h.file <- system.file("extdata/input", "1000GP_Phase3_subset_chr10.hap.gz", package = "baseqtl", mustWork = TRUE)
AI_estimate <- system.file("extdata/input", "AI_estimate.psoriasis.txt", package = "baseqtl", mustWork = TRUE)

## Choose your output directory
out <- tempdir()

## To minimise file sizes and computational load SNPs are within a 10^4 cis-window
## baseqtl is computational intense, it is recommended to be run with several cores
## cores are automatically detected by R
## Run baseqtl.n2Togt:
\dontrun{
baseqtl2T.nogt(
  gene = "ENSG00000178372",
  chr = 10,
  snps = 10^3,
  counts.f = counts.f,
  covariates = covariates,
  e.snps = e.snps,
  u.esnps = u.snps,
  gene.coord = gene.coord,
  vcf = vcf,
  sample.f = sample.f,
  le.file = le.file,
  h.file = h.file,
  out = out,
  treatment = c("Psoriasis_skin", "normal_skin"),
  AI_estimate = AI_estimate
)
}
}
